small mods to syntax

+++ src/main/resources/db/migration/V1__create_initial_tables.sql
@@ -1,7 +1,7 @@
 CREATE TABLE food_chains (
     id SERIAL PRIMARY KEY,
     name VARCHAR NOT NULL,
-    rating INTEGER,
+    rating NUMERIC(2,1),
     delivery BOOLEAN NOT NULL,
     description VARCHAR(255),
     img_url TEXT NOT NULL
